requirements/base.txt
# SHA1:0862095245a068ae2fc00217da78331e1e7ae505
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-e file:.
    # via -r requirements/base.in
aiohttp==3.7.2
    # via slackclient
alembic==1.4.3
    # via flask-migrate
amqp==2.6.1
    # via kombu
apispec[yaml]==3.3.2
    # via flask-appbuilder
async-timeout==3.0.1
    # via aiohttp
attrs==20.2.0
    # via
    #   aiohttp
    #   jsonschema
babel==2.8.0
    # via flask-babel
backoff==1.10.0
    # via apache-rabbitai
billiard==3.6.3.0
    # via celery
bleach==3.2.1
    # via apache-rabbitai
brotli==1.0.9
    # via flask-compress
cachelib==0.1.1
    # via apache-rabbitai
celery==4.4.7
    # via apache-rabbitai
cffi==1.14.3
    # via cryptography
chardet==3.0.4
    # via aiohttp
click==7.1.2
    # via
    #   apache-rabbitai
    #   flask
    #   flask-appbuilder
colorama==0.4.4
    # via
    #   apache-rabbitai
    #   flask-appbuilder
contextlib2==0.6.0.post1
    # via apache-rabbitai
convertdate==2.3.0
    # via holidays
cron-descriptor==1.2.24
    # via apache-rabbitai
croniter==0.3.36
    # via apache-rabbitai
cryptography==3.2.1
    # via apache-rabbitai
decorator==4.4.2
    # via retry
defusedxml==0.6.0
    # via python3-openid
dnspython==2.0.0
    # via email-validator
email-validator==1.1.1
    # via flask-appbuilder
flask-appbuilder==3.3.0
    # via apache-rabbitai
flask-babel==1.0.0
    # via flask-appbuilder
flask-caching==1.9.0
    # via apache-rabbitai
flask-compress==1.8.0
    # via apache-rabbitai
flask-jwt-extended==3.24.1
    # via flask-appbuilder
flask-login==0.4.1
    # via flask-appbuilder
flask-migrate==2.5.3
    # via apache-rabbitai
flask-openid==1.2.5
    # via flask-appbuilder
flask-sqlalchemy==2.4.4
    # via
    #   flask-appbuilder
    #   flask-migrate
flask-talisman==0.7.0
    # via apache-rabbitai
flask-wtf==0.14.3
    # via
    #   apache-rabbitai
    #   flask-appbuilder
flask==1.1.2
    # via
    #   apache-rabbitai
    #   flask-appbuilder
    #   flask-babel
    #   flask-caching
    #   flask-compress
    #   flask-jwt-extended
    #   flask-login
    #   flask-migrate
    #   flask-openid
    #   flask-sqlalchemy
    #   flask-wtf
geographiclib==1.50
    # via geopy
geopy==2.0.0
    # via apache-rabbitai
graphlib-backport==1.0.3
    # via apache-rabbitai
gunicorn==20.0.4
    # via apache-rabbitai
holidays==0.10.3
    # via apache-rabbitai
humanize==3.1.0
    # via apache-rabbitai
idna==2.10
    # via
    #   email-validator
    #   yarl
importlib-metadata==2.1.1
    # via
    #   jsonschema
    #   kombu
    #   markdown
isodate==0.6.0
    # via apache-rabbitai
itsdangerous==1.1.0
    # via
    #   apache-rabbitai
    #   flask
    #   flask-wtf
jinja2==2.11.3
    # via
    #   flask
    #   flask-babel
jsonschema==3.2.0
    # via flask-appbuilder
kombu==4.6.11
    # via celery
korean-lunar-calendar==0.2.1
    # via holidays
mako==1.1.3
    # via alembic
markdown==3.3.3
    # via apache-rabbitai
markupsafe==1.1.1
    # via
    #   jinja2
    #   mako
    #   wtforms
marshmallow-enum==1.5.1
    # via flask-appbuilder
marshmallow-sqlalchemy==0.23.1
    # via flask-appbuilder
marshmallow==3.9.0
    # via
    #   flask-appbuilder
    #   marshmallow-enum
    #   marshmallow-sqlalchemy
msgpack==1.0.0
    # via apache-rabbitai
multidict==5.0.0
    # via
    #   aiohttp
    #   yarl
natsort==7.0.1
    # via croniter
numpy==1.19.4
    # via
    #   pandas
    #   pyarrow
packaging==20.4
    # via bleach
pandas==1.2.2
    # via apache-rabbitai
parsedatetime==2.6
    # via apache-rabbitai
pathlib2==2.3.5
    # via apache-rabbitai
pgsanity==0.2.9
    # via apache-rabbitai
polyline==1.4.0
    # via apache-rabbitai
prison==0.1.3
    # via flask-appbuilder
py==1.9.0
    # via retry
pyarrow==3.0.0
    # via apache-rabbitai
pycparser==2.20
    # via cffi
pyjwt==1.7.1
    # via
    #   apache-rabbitai
    #   flask-appbuilder
    #   flask-jwt-extended
pymeeus==0.3.7
    # via convertdate
pyparsing==2.4.7
    # via
    #   apache-rabbitai
    #   packaging
pyrsistent==0.16.1
    # via
    #   -r requirements/base.in
    #   jsonschema
python-dateutil==2.8.1
    # via
    #   alembic
    #   apache-rabbitai
    #   croniter
    #   flask-appbuilder
    #   holidays
    #   pandas
python-dotenv==0.15.0
    # via apache-rabbitai
python-editor==1.0.4
    # via alembic
python-geohash==0.8.5
    # via apache-rabbitai
python3-openid==3.2.0
    # via flask-openid
pytz==2020.4
    # via
    #   babel
    #   celery
    #   convertdate
    #   flask-babel
    #   pandas
pyyaml==5.4.1
    # via
    #   apache-rabbitai
    #   apispec
redis==3.5.3
    # via apache-rabbitai
retry==0.9.2
    # via apache-rabbitai
selenium==3.141.0
    # via apache-rabbitai
simplejson==3.17.2
    # via apache-rabbitai
six==1.15.0
    # via
    #   bleach
    #   cryptography
    #   flask-jwt-extended
    #   flask-talisman
    #   holidays
    #   isodate
    #   jsonschema
    #   packaging
    #   pathlib2
    #   polyline
    #   prison
    #   pyrsistent
    #   python-dateutil
    #   sqlalchemy-utils
    #   wtforms-json
slackclient==2.5.0
    # via apache-rabbitai
sqlalchemy-utils==0.36.8
    # via
    #   apache-rabbitai
    #   flask-appbuilder
sqlalchemy==1.3.20
    # via
    #   alembic
    #   apache-rabbitai
    #   flask-appbuilder
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
    #   sqlalchemy-utils
sqlparse==0.3.0
    # via apache-rabbitai
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   apache-rabbitai
    #   yarl
urllib3==1.25.11
    # via selenium
vine==1.3.0
    # via
    #   amqp
    #   celery
webencodings==0.5.1
    # via bleach
werkzeug==1.0.1
    # via
    #   flask
    #   flask-jwt-extended
wtforms-json==0.3.3
    # via apache-rabbitai
wtforms==2.3.3
    # via
    #   flask-wtf
    #   wtforms-json
yarl==1.6.2
    # via aiohttp
zipp==3.4.1
    # via
    #   -r requirements/base.in
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
